+++ gear/form_qc_coordinator/src/python/form_qc_coordinator_app/coordinator.py
@@ -13,6 +13,7 @@ from flywheel_adaptor.subject_adaptor import SubjectAdaptor, VisitInfo
 from flywheel_gear_toolkit import GearToolkitContext
 from flywheel_gear_toolkit.utils.metadata import Metadata, create_qc_result_dict
 from gear_execution.gear_execution import GearExecutionError
+from gear_execution.gear_trigger import GearConfigs, GearInfo
 from keys.keys import FieldNames
 from outputs.errors import (
     FileError,
@@ -22,16 +23,12 @@ from outputs.errors import (
     system_error,
     update_error_log_and_qc_metadata,
 )
-from pydantic import BaseModel, ConfigDict
 
 log = logging.getLogger(__name__)
 
 
-class QCGearConfigs(BaseModel):
+class QCGearConfigs(GearConfigs):
     """Class to represent qc gear configs."""
-    model_config = ConfigDict(populate_by_name=True)
-
-    apikey_path_prefix: str
     rules_s3_bucket: str
     qc_checks_db_path: str
     primary_key: str
@@ -39,13 +36,6 @@ class QCGearConfigs(BaseModel):
     strict_mode: Optional[bool] = True
     legacy_project_label: Optional[str] = None
     date_field: Optional[str] = None
-
-
-class QCGearInfo(BaseModel):
-    model_config = ConfigDict(populate_by_name=True)
-
-    gear_name: str
-    configs: QCGearConfigs
 
 
 class QCCoordinator():
@@ -208,29 +198,21 @@ class QCCoordinator():
                                visitdate=visitdate)
         self.__subject.set_last_failed_visit(self.__module, visit_info)
 
-    def run_error_checks(self, *, gear_name: str, gear_configs: QCGearConfigs,
+    def run_error_checks(self, *, qc_gear_info: GearInfo,
                          visits: List[Dict[str, str]], date_col: str) -> None:
         """Sequentially trigger the QC checks gear on the provided visits. If a
         visit failed QC validation or error occurred while running the QC gear,
         none of the subsequent visits will be evaluated.
 
         Args:
-            gear_name: QC checks gear name (form_qc_checker)
-            gear_configs: QC check gear configs
+            qc_gear_info: GearInfo containing info for the qc gear
             visits: set of visits to be evaluated
             date_col: name of the visit date field to sort the visits
 
         Raises:
             GearExecutionError if errors occur while triggering the QC gear
         """
-
-        try:
-            gear = self.__proxy.lookup_gear(gear_name)
-        except ApiException as error:
-            raise GearExecutionError(error) from error
-
-        configs = gear_configs.model_dump()
-
+        gear_name = qc_gear_info.gear_name
         ptid_key = f'file.info.forms.json.{FieldNames.PTID}'
         date_col_key = f'file.info.forms.json.{date_col}'
 
@@ -254,9 +236,9 @@ class QCCoordinator():
                 raise GearExecutionError(
                     f'Failed to retrieve {filename} - {error}') from error
 
-            job_id = gear.run(config=configs,
-                              inputs={"form_data_file": visit_file},
-                              destination=destination)
+            job_id = qc_gear_info.trigger_gear(proxy=self.__proxy,
+                                               inputs={"form_data_file": visit_file},
+                                               destination=destination)
             if job_id:
                 log.info('Gear %s queued for file %s - Job ID %s', gear_name,
                          filename, job_id)