+++ gear/prescreening/src/python/prescreening_app/main.py
@@ -1,5 +1,6 @@
 """Defines Prescreening."""
 import logging
+from flywheel.models.job_state import JobState
 from typing import List
 
 from flywheel_adaptor.flywheel_proxy import FlywheelProxy
@@ -7,16 +8,16 @@ from gear_execution.gear_execution import (
     GearExecutionError,
     InputFileWrapper,
 )
+from gear_execution.gear_trigger import GearInfo
 
 log = logging.getLogger(__name__)
-
 
 def run(*,
         proxy: FlywheelProxy,
         file_input: InputFileWrapper,
         accepted_modules: List[str],
         tags_to_add: List[str],
-        local_run: bool = False) -> None:
+        scheduler_gear: GearInfo) -> None:
     """Runs the prescreening process. Checks that the file suffix matches any
     accepted modules; if so, tag the file with the specified tags, else report
     error.
@@ -27,16 +28,15 @@ def run(*,
             potentially queue
         accepted_modules: List of accepted modules (case-insensitive)
         tags_to_add: List of tags to add if the file passes prescreening
-        local_run: Whether or not this is a local run - if so will verify the
-            file but won't push changes upstream
+        scheduler_gear: GearInfo of the scheduler gear to trigger
     """
     module = file_input.basename.split('-')[-1]
     if module.lower() not in accepted_modules:
         raise GearExecutionError(f"Unallowed module suffix: {module}")
 
-    if local_run or proxy.dry_run:
-        log.info("Dry run or local run set: file passes prescreening, " +
-                 f"would have added {tags_to_add}")
+    if proxy.dry_run:
+        log.info("DRY RUN: file passes prescreening, would have added" +
+                 f"{tags_to_add} and triggerd {scheduler_gear}")
         return
 
     # add the specified tag
@@ -44,3 +44,17 @@ def run(*,
     file = proxy.get_file(file_input.file_id)
     for tag in tags_to_add:
         file.add_tag(tag)
+
+    # check if the scheduler gear is pending/running
+    project_id = file.file_parents.project
+    states = [JobState.RUNNING, JobState.PENDING]
+    log.info(f"Checking status of {scheduler_gear.gear_name}")
+    if scheduler_gear.check_instance_by_state(proxy=proxy,
+                                              states=states,
+                                              project_id=project_id):
+        log.info("Scheduler gear already running")
+        return
+
+    # otherwise invoke the gear
+    scheduler_gear.run(proxy=proxy,
+                       destination=proxy.get_project_by_id(project_id))