+++ docs/prescreening/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 All notable changes to this gear are documented in this file.
 
-## 0.1.0
+## 0.0.3
 
 * Initial version
+* Adds the `gear_execution/gear_trigger` common code to handle triggering of gears
 * Adds this CHANGELOG